services/quant_engine/requirements.txt
uvloop==0.19.0
pyzmq==25.1.2
redis[asyncio]==5.0.1
orjson==3.9.15
python-dotenv==1.0.1
